ultralytics-main-django/requirements.txt
dill==0.3.9
Django==4.2.16
djangorestframework==3.15.2
PyMySQL==1.1.1
Requests==2.32.3
seaborn==0.13.2
django-cors-headers
channels